Converting ValenciaMECq0q3ResponseCalc into header only
