+++ src/nusystematics/responsecalculators/CMakeLists.txt
@@ -18,6 +18,7 @@ SET(RSPC_HDRFILES
   DIRT2_EmissEngine_Reweight.hh
   CCQERPAReweightCalculator.hh
   FSIReweightCalculator.hh
+  ValenciaMECq0q3ResponseCalc.hh
 )
